fix: Uncached data was accessed outside of <Suspense>
